move ih logger functionality to ih analytics adapter
